Alpha-0.0.23.0 Implemented AbilityDescription struct Ability no longer implements IFlagged<T>

+++ DMTools/Items.cs
@@ -250,7 +250,7 @@ namespace DnDTDesktop
         {
             get
             {
-                return Flags[FlagList.QuantityLocked] ? Convert.ToInt32(quant) : quant;
+                return Flags[FlagList.QuantityLocked] ? Convert.ToUInt32(quant) : (uint)quant;
             }
             set
             {
